move helpers around; simplify fixture
